+++ crates/kct_package/src/compiler/property.rs
@@ -1,11 +1,22 @@
+use crate::compiler::Compiler;
+
 use serde_json::Value;
-use std::hash::Hash;
+use std::{collections::HashMap, hash::Hash};
+
+type Handler = Box<dyn Fn(HashMap<String, Value>) -> Result<Value, String> + 'static>;
+
+pub struct Function {
+	pub params: Vec<String>,
+	pub handler: Handler,
+}
 
 #[derive(Clone, Hash, PartialEq, Eq)]
 pub enum Name {
+	File,
+	Include,
+	Input,
 	Package,
 	Release,
-	Input,
 }
 
 impl Name {
@@ -13,6 +24,8 @@ impl Name {
 		use Name::*;
 
 		match self {
+			File => "files",
+			Include => "include",
 			Package => "package",
 			Release => "release",
 			Input => "input",
@@ -20,8 +33,13 @@ impl Name {
 	}
 }
 
+pub enum Output {
+	Plain(Value),
+	Callback(Function),
+}
+
 pub trait Property {
 	fn name(&self) -> Name;
 
-	fn generate(&self) -> Value;
+	fn generate(&self, compiler: &Compiler) -> Output;
 }