+++ crates/kct_package/src/lib.rs
@@ -1,8 +1,11 @@
 mod archiver;
 mod compiler;
 mod error;
+mod functions;
+mod input;
 mod package;
+mod release;
 
-pub use crate::compiler::Release;
 pub use crate::error::Error;
 pub use crate::package::Package;
+pub use crate::release::Release;